Working towards the integration of robosuite with https://github.com/learningLogisticsLab/rlkit-relational.
1) To get this running,  you will need to install both robosuite and rlkit-relational.

The integration is yet incomplete, but in this commit we have a number of modifications. See descriptions listed below:

- load_objects_to_simulate(): extensive redoing here. Tries to be smart about loading objects from file and check for different naming conventions to avoid making mistakes. TODO: modifiy assumption of oXXXXv naming condition if we change to 'VisualObject' format.

- reset_internal(): it is in this method where we will select a strategy to place objects. It has been upgraded significantly. Also has assumptions for naming conventions here as oXXXXv. TODO: the eefSampler was created to set a eef pose. But to actually move the robot there, we either have to call an IK method to get the desired joint positions and then set them on mujoco and move the simulation forward, or alternatively create an interpolated joint trajectory from desired to actual, and step forward to hvae the controller move us there.

1. Bins or bins_arena
Changed the target bin by removing sub-compartments and make it just a large bin as the source bin # 1. This replicated the bin picking environments for induction and also allows to work with more objects. May need to change bin dimensions in the future.

2. Objects
The robosuite/models/objects/__init__.py lists all object classes we will use. We assume objects will be called oXXXX and oXXXXv. This may be modified again to match robosuites assumptions of ...Object and ...VisualObject nomenclatures.

3. Environments
The core part of the commit is the adaptation of the environments/manipulation/picking.py script. Follows the same format as pick_place.py which is standard in robosuite.

Still incomplete, but roughly these are the modifications made so far:
- load_model(): only aesthetic. calls get_placement_initializer.

- get_placement_initializer(): changed the names of the placement_ininitializers to better reflect (the random setting of pose for) picking, placing, and created a new sampler to randomly set the robot's eef (is used in load_model() but also in _reset_internals()).

- setup_references(): aesthetic. TODO: update bin_x_low, bin_y_low for self.target_bin_placements.

- create observables(): aesthetic. TODO: may add observables to match the original observations used in relational_RL, which build on fetch tasks.

- create_obj_sensors(): no changes planned for now.

- visualize(): no changes in picking.py but we do activate the visualization when in base.py:MujocoEnv.reset() for debugging purposes.

Still to be updated/discarded:
- reward()
- staged_rewards()
- not_in_bin()
- check_suchess()