In this commit, we worked on appropriately setting the goal throughout the environment.
At each step, only one object of those availables is selected to be the goal object to pick.
Remember this is encoded into the DRL algoright by only considering the rel pose between gripper and this object in our observations.

The goal object is a dict with 'name, 'pos', and 'quat' keys and stored in self.goal_object.
It is acquired by get_goal_boject() which randomly selects a name and gets the goal pos/quat from
the self.object_placements dictionary (created in reset_internals()). get_goal_object() is called immediately after setting the placements in reset_internals().

Additionally, the goal object is also changed when we succeed in placing it. This is done in is_success(). Note that is_success() name was set like this from _check_success() to match relationalRL's interface. is_success() tests whether the pos (not orientation yet), is within some tolerance of the goal_pos. If so, remove this object from the list of objects considered for the next round.

Finally, this goal's pos and quat is used as the 'desired_goal' that we have in _get_obs() as part of the HER algorithm.